Changes made while debugging docker files.
