+++ docker-compose.yml
@@ -4,8 +4,6 @@
 # Place this in the project root directory
 # ============================================
 
-version: '3.8'
-
 services:
   # PostgreSQL Database
   db:
@@ -16,7 +14,7 @@ services:
       POSTGRES_USER: postgres
       POSTGRES_PASSWORD: ${DB_PASSWORD:-changeme}  # Use env var or default
     ports:
-      - "5432:5432"
+      - "5433:5432"
     volumes:
       - volunteer-db-data:/var/lib/postgresql/data
     networks:
@@ -30,7 +28,7 @@ services:
 
   # Go GraphQL API
   api:
-    build: ./api
+    build: ./vol_sched_api
     container_name: volunteer-api
     environment:
       DATABASE_URL: postgres://postgres:${DB_PASSWORD:-changeme}@db:5432/volunteer_scheduler?sslmode=disable
@@ -51,7 +49,7 @@ services:
 
   # Next.js Frontend
   frontend:
-    build: ./frontend
+    build: ./vol_sched_app
     container_name: volunteer-frontend
     environment:
       NEXT_PUBLIC_API_URL: http://localhost:8080/query
