+++ vol_sched_api/Dockerfile
@@ -1,13 +1,14 @@
 # ============================================
-# api/Dockerfile
+# vol_sched_api/Dockerfile
 # Go GraphQL API for Volunteer Scheduler
 # ============================================
 
 # Build stage
 FROM golang:1.25-alpine AS builder
 
-# Install build dependencies
+# Install build dependencies and gqlgen
 RUN apk add --no-cache git
+RUN go install github.com/99designs/gqlgen@latest
 
 # Set working directory
 WORKDIR /app
@@ -21,8 +22,11 @@ RUN go mod download
 # Copy source code
 COPY . .
 
+# Generate GraphQL code
+RUN gqlgen generate
+
 # Build the application
-RUN CGO_ENABLED=0 GOOS=linux go build -a -installsuffix cgo -o volunteer-api server.go
+RUN CGO_ENABLED=0 GOOS=linux go build -a -installsuffix cgo -o volunteer-api .
 
 # Final stage
 FROM alpine:latest
@@ -35,9 +39,8 @@ WORKDIR /root/
 # Copy the binary from builder
 COPY --from=builder /app/volunteer-api .
 
-# Copy schema and config files
+# Copy schema files
 COPY --from=builder /app/schema.graphql .
-COPY --from=builder /app/gqlgen.yml .
 
 # Expose port
 EXPOSE 8080
